Fixes major NMT metric tracking bug.
When we implemented this originally, we didn't fully think through how
metric sets were used by DropWizard.  The gauges are registered only
once, and it's when the gauge function is _called_ that you have an
opportunity to get the latest value of interest.

This new implementation also adds support for the ever-expanding set of
NMT metrics that show up when you query the VM.