Removes memory free rate estimation.
Jira: OTPL-1895

It didn't really work as desired.  The output was wildly varying, and
not really usable as the basis for an alert.  The GC memory metrics look
like they have some real potential.  So this commit pairs down the
memory free metrics to contribute *only* the free gauge.  Further, this
commit updates the namespace in which it's emitted to be the same as
that of the Dropwizard memory usage gauge set.  This means the two
namespaces can be combined for easier use.